Selesai

+++ index.php
@@ -1,14 +1,14 @@
 <?php 
-include 'koneksi.php';
+	include 'koneksi.php';
 
-$sql = "SELECT id, nama_depan, nama_belakang, jenis_kelamin, hp_orangtua, kelas, alamat FROM santri";
-$result = mysqli_query($connect, $sql);
+	$sql = "SELECT id, nama_depan, nama_belakang, jenis_kelamin, hp_orangtua, kelas, alamat FROM santri";
 
-$data_santri = mysqli_fetch_all($result, MYSQLI_ASSOC);
+	$result = mysqli_query($connect, $sql);
 
-mysqli_free_result($result);
-mysqli_close($connect);
+	$data_santri = mysqli_fetch_all($result, MYSQLI_ASSOC);
 
+	mysqli_free_result($result);
+	mysqli_close($connect);
 ?>
 <!DOCTYPE html>
 <html>
@@ -32,23 +32,24 @@ mysqli_close($connect);
 		    <th colspan="2">Aksi</th>
 		</tr>
 		<?php 
-			if (empty($data_santri)) {
+			if(empty($data_santri)) {
 				echo '<tr>
-			  		<td colspan="8" style="text-align: center;">Belum ada data santri.</td>
-			  	</tr>';
-			} else {
-			  	foreach($data_santri as $santri) {
-			  		echo '<tr>
-					    <td>' . $santri['nama_depan'] . '</td>
-					    <td>' . $santri['nama_belakang'] . '</td>
-					    <td>' . $santri['jenis_kelamin'] . '</td>
-					    <td>' . $santri['hp_orangtua'] . '</td>
-					    <td>' . $santri['kelas'] . '</td>
-					    <td>' . $santri['alamat'] . '</td>
-					    <td><a href="edit.php?id=' . $santri['id'] . '">Edit</a></td>
-					    <td><a href="hapus.php?id=' . $santri['id'] . '">Hapus</a></td>
+					<td colspan="8">Belum ada data santri.</td>
+				</tr>';
+			}
+			else {
+				foreach ($data_santri as $santri) {
+					echo '<tr>
+						<td>' . $santri['nama_depan'] . '</td>
+						<td>' . $santri['nama_belakang'] . '</td>
+						<td>' . $santri['jenis_kelamin'] . '</td>
+						<td>' . $santri['hp_orangtua'] . '</td>
+						<td>' . $santri['kelas'] . '</td>
+						<td>' . $santri['alamat'] . '</td>
+						<td><a href="edit.php?id=' . $santri['id'] .'">Edit</a></td>
+						<td><a onclick="return confirm(\'Apakah anda yakin menghapus data ini?\')" href="hapus.php?id=' . $santri['id'] .'" >Hapus</a></td>
 					</tr>';
-			  	}
+				}
 			}
 		?>
 	</table>
